+++ packages/llm-bot/src/core/OptimisticOracleV2.ts
