+++ packages/llm-bot/src/core/common.ts
