+++ packages/llm-bot/src/core/examples.ts
