+++ packages/llm-bot/src/utils/contracts.ts
